Show and update focus line in all graphs
The focus line of each graph now follows the movements in one graph
through a notifier pattern. The focus line position is updated such that
one can compare graph values more easily. Add transparency to the focus
circle, and let it follow the actual value line if possible in the
component graphs.